requirements.txt
aiogram
# aiohttp==4.0.0a1
# aioredis==1.3.1
python-dotenv
SQLAlchemy
asyncpg  # для работы Алхимии

# selenium==4.0.0a7
beautifulsoup4
# requests==2.25.1
# gino
vkbottle
httpx

# FIXME: Не получается обновить pip, какая-то ошибка ENVS